[chore]: Eliminate cli namespace and rename common support header to utility
