+++ lib/compiler/include/sema/type.hpp
@@ -3,12 +3,11 @@
 #include <cassert>
 #include <span>
 #include <type_traits>
-#include <utility>
 
 #include <ankerl/unordered_dense.h>
 
 #include "array.hpp"
-#include "common.hpp"
+#include "utility.hpp"
 #include "hash.hpp"
 #include "optional.hpp"
 #include "types.hpp"
