+++ lib/compiler/tests/helpers/ast.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <algorithm>
-#include <utility>
 #include <vector>
 
 #include <catch2/catch_test_macros.hpp>
