+++ lib/driver/include/platform/win32.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-namespace porpoise::cli::win32 {
+namespace porpoise::win32 {
 
 // Enables UTF8 on creation, disables on destruction
 //
@@ -13,4 +13,4 @@ class RichConsole {
 #endif
 };
 
-} // namespace porpoise::cli::win32
+} // namespace porpoise::win32
